+++ src/Registry/TypeRegistry.php
@@ -49,6 +49,7 @@ class TypeRegistry implements Registrable {
 		// Enums to register.
 		$classes_to_register = [
 			Enum\ProximityRadiusOptions::class,
+			Enum\FacetTypeEnum::class,
 			Enum\SortOptionsEnum::class,
 		];
 
@@ -69,6 +70,7 @@ class TypeRegistry implements Registrable {
 	 */
 	private static function inputs(): array {
 		$classes_to_register = [
+			Input\FacetsInput::class,
 			Input\DateRangeArgs::class,
 			Input\NumberRangeArgs::class,
 			Input\ProximityArgs::class,
@@ -93,6 +95,7 @@ class TypeRegistry implements Registrable {
 	private static function interfaces(): array {
 		$classes_to_register = [
 			WPInterface\FacetConfig::class,
+			WPInterface\NodeWithFacets::class,
 		];
 
 		/**
@@ -119,6 +122,7 @@ class TypeRegistry implements Registrable {
 			WPObject\FacetSortOptionSetting::class,
 			WPObject\FacetSettings::class,
 			WPObject\Facet::class,
+			WPObject\FacetObjects::class,
 		];
 
 		/**